Safari-611.1.10/LayoutTests/webgpu/map-write-buffers-expected.txt

CONSOLE MESSAGE: GPUBuffer.mapWriteAsync(): Invalid operation!
CONSOLE MESSAGE: GPUBuffer.mapWriteAsync(): Invalid operation!

PASS Map write, unmap, and destroy a GPUBuffer.
PASS Reject a map write on a buffer not created with MAP_WRITE usage.
PASS Reject a map write on a mapped GPUBuffer.
PASS Reject a pending map write if GPUBuffer is unmapped.
